backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
bidict==0.23.1
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
ecdsa==0.19.1
fastapi==0.115.14
greenlet==3.2.3
h11==0.16.0
idna==3.10
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
python-engineio==4.12.2
python-jose==3.5.0
python-multipart==0.0.20
python-socketio==5.13.0
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
wsproto==1.2.0
